Moving a project to visual studio 2019
