--- CProject/Project1/Shared.h
@@ -1,3 +1,0 @@
-#include "stdint.h"
-#include "Windows.h"
-#include <iostream>
