--- CProject/Sockets.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.1022
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Project1", "Project1\Project1.vcxproj", "{B41DF057-14F8-4D95-B247-E8F007814FE5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B41DF057-14F8-4D95-B247-E8F007814FE5}.Debug|x64.ActiveCfg = Debug|x64
-		{B41DF057-14F8-4D95-B247-E8F007814FE5}.Debug|x64.Build.0 = Debug|x64
-		{B41DF057-14F8-4D95-B247-E8F007814FE5}.Debug|x86.ActiveCfg = Debug|Win32
-		{B41DF057-14F8-4D95-B247-E8F007814FE5}.Debug|x86.Build.0 = Debug|Win32
-		{B41DF057-14F8-4D95-B247-E8F007814FE5}.Release|x64.ActiveCfg = Release|x64
-		{B41DF057-14F8-4D95-B247-E8F007814FE5}.Release|x64.Build.0 = Release|x64
-		{B41DF057-14F8-4D95-B247-E8F007814FE5}.Release|x86.ActiveCfg = Release|Win32
-		{B41DF057-14F8-4D95-B247-E8F007814FE5}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {AFA6A525-9C8C-43ED-8F67-724CBF20A3AA}
-	EndGlobalSection
-EndGlobal
