+++ Sockets/Sockets.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30225.117
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Sockets", "Sockets\Sockets.vcxproj", "{4EA82211-630A-4530-AEFA-29603EB6214F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4EA82211-630A-4530-AEFA-29603EB6214F}.Debug|x64.ActiveCfg = Debug|x64
+		{4EA82211-630A-4530-AEFA-29603EB6214F}.Debug|x64.Build.0 = Debug|x64
+		{4EA82211-630A-4530-AEFA-29603EB6214F}.Debug|x86.ActiveCfg = Debug|Win32
+		{4EA82211-630A-4530-AEFA-29603EB6214F}.Debug|x86.Build.0 = Debug|Win32
+		{4EA82211-630A-4530-AEFA-29603EB6214F}.Release|x64.ActiveCfg = Release|x64
+		{4EA82211-630A-4530-AEFA-29603EB6214F}.Release|x64.Build.0 = Release|x64
+		{4EA82211-630A-4530-AEFA-29603EB6214F}.Release|x86.ActiveCfg = Release|Win32
+		{4EA82211-630A-4530-AEFA-29603EB6214F}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E36C06DF-4061-49E3-90BA-0652DB0A7380}
+	EndGlobalSection
+EndGlobal
