+++ Sockets/Sockets/Shared.h
@@ -1,0 +1,3 @@
+#include "stdint.h"
+#include "Windows.h"
+#include <iostream>
